requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
sqlalchemy<2.0.0
flask-login==0.6.2
werkzeug==2.3.7
torch==2.2.1
torchvision==0.17.1
pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80
nibabel==5.2.0
scipy==1.12.0
reportlab==4.1.0
pydicom==2.4.4